Problem: tree sum

+++ 027-tree-sum/cases_test.go
@@ -7,7 +7,7 @@ import (
 type testCase struct {
 	name     string
 	root     *treesum.Node
-	expected int64
+	expected int
 }
 
 //nolint:gochecknoglobals
