+++ 027-tree-sum/node.go
@@ -1,7 +1,7 @@
 package treesum
 
 type Node struct {
-	Val   int64
+	Val   int
 	Left  *Node
 	Right *Node
 }
